# docs/public/api/openapi-3.0.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/entities": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Query All Entities",
                "parameters": [
                    {
                        "description": "search string",
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "items per page",
                        "name": "pageSize",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "tags Ids",
                        "name": "tags",
                        "in": "query",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "description": "parent Ids",
                        "name": "parentIds",
                        "in": "query",
                        "explode": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.EntityListResult"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Create Entity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/repo.EntityCreate"
                            }
                        }
                    },
                    "description": "Entity Data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.EntityOut"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/export": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Export Entities",
                "responses": {
                    "200": {
                        "description": "text/csv",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/fields": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/fields/values": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Values",
                "parameters": [
                    {
                        "description": "Field name",
                        "name": "field",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/import": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Import Entities",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "csv": {
                                        "description": "CSV file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "csv"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/entities/tree": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get Locations Tree",
                "parameters": [
                    {
                        "description": "include items in response tree",
                        "name": "withItems",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/repo.TreeItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/entities/{id}/duplicate": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Duplicate Entity",
                "parameters": [
                    {
                        "description": "Entity ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/repo.DuplicateOptions"
                            }
                        }
                    },
                    "description": "Duplicate Options",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/repo.EntityOut"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/entities/{id}/path": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get the full path of an entity",
                "parameters": [
                    {
                        "description": "Entity ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/repo.EntityPath"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "repo.DuplicateOptions": {
                "type": "object",
                "properties": {
                    "copyAttachments": {
                        "type": "boolean"
                    },
                    "copyCustomFields": {
                        "type": "boolean"
                    },
                    "copyMaintenance": {
                        "type": "boolean"
                    },
                    "copyPrefix": {
                        "type": "string"
                    }
                }
            },
            "repo.EntityCreate": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "description": {
                        "type": "string",
                        "maxLength": 1000
                    },
                    "entityTypeId": {
                        "type": "string"
                    },
                    "manufacturer": {
                        "type": "string",
                        "maxLength": 255,
                        "x-omitempty": true,
                        "nullable": true
                    },
                    "modelNumber": {
                        "description": "Identifications — optional at create time; populated e.g. by the\nbarcode product-search import flow (#1578).",
                        "type": "string",
                        "maxLength": 255,
                        "x-omitempty": true,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1
                    },
                    "parentId": {
                        "type": "string",
                        "nullable": true
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "tagIds": {
                        "description": "Edges",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "repo.EntityListResult": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/repo.EntitySummary"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "pageSize": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "totalPrice": {
                        "type": "number"
                    }
                }
            },
...
            "repo.TreeItem": {
                "type": "object",
                "properties": {
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/repo.TreeItem"
                        }
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}